refactor: clean modules and move generic type
