+++ service/src/common/mod.rs
@@ -4,5 +4,4 @@
 pub mod address;
 pub mod database;
 pub mod indexer_error;
-// pub mod query_fee_models;
-// pub mod schema;
+pub mod types;
