--- service/src/common/schema.rs
@@ -1,16 +1,0 @@
-// Copyright 2023-, GraphOps and Semiotic Labs.
-// SPDX-License-Identifier: Apache-2.0
-
-// diesel::table! {
-//     use diesel::sql_types::*;
-//     use crate::enums::diesel_exports::*;
-
-//     allocation_receipts (id) {
-//         id -> Varchar,
-//         allocation -> Varchar,
-//         fees -> Varchar,
-//         signature -> Varchar,
-//         created_at -> Timestamp,
-//         updated_at -> Timestamp,
-//     }
-// }
