--- service/src/common/verifier.rs
@@ -1,3 +1,0 @@
-// Copyright 2023-, GraphOps and Semiotic Labs.
-// SPDX-License-Identifier: Apache-2.0
-
